servicebinding: fix a bug in ownership check
Before this change, the reconciler would check the
new CRD to see if it had the correct owner reference.
But the new CRD always has the correct ownership reference!

After this change, the reconciler checks if the existing
CRD to see if it has the correct owner reference, which
seems like the intended behavior.

Signed-off-by: Nick Santos <nick.santos@docker.com>